+++ src/app/app.module.ts
@@ -1,23 +1,14 @@
-import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 
 import { AppRoutingModule } from "./app-routing.module";
 import { AppComponent } from "./app.component";
-import { ProductComponent, ProductListComponent, ProductService } from "./products/index";
-import { CartComponent, CartService } from "./cart/index";
+import { ProductsModule } from "./products/index";
+import { CartModule } from "./cart/index";
+import { SharedModule } from "./shared/index";
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule
-  ],
-  providers: [ProductService, CartService],
-  bootstrap: [AppComponent]
+  declarations: [AppComponent],
+  imports: [AppRoutingModule, CartModule, ProductsModule, SharedModule],
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}