+++ src/app/cart/components/cart-item/cart-item.component.spec.ts
@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { CartComponent } from "./cart.component";
+import { CartItemComponent } from "./cart-item.component";
 
-describe("CartComponent", () => {
-  let component: CartComponent;
-  let fixture: ComponentFixture<CartComponent>;
+describe("CartItemComponent", () => {
+  let component: CartItemComponent;
+  let fixture: ComponentFixture<CartItemComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CartComponent ]
+      declarations: [ CartItemComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CartComponent);
+    fixture = TestBed.createComponent(CartItemComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });