+++ src/app/cart/components/cart-list/index.ts
@@ -1,0 +1,1 @@
+export { CartListComponent } from "./cart-list.component";