+++ src/app/core/models/category-id.model.ts
@@ -1,7 +1,0 @@
-/**
- * Category Id model
- */
-export interface CategoryId {
-  name: string;
-  isVisible?: boolean;
-}