+++ src/app/products/components/product/product.component.ts
@@ -1,31 +1,33 @@
-import { Component, OnInit, ChangeDetectionStrategy, Output, EventEmitter } from "@angular/core";
+import { Component, ChangeDetectionStrategy, Output, EventEmitter, Input } from "@angular/core";
 
+import { ProductModel } from "../../models/index";
+
+/**
+ * Smart component that represents product list
+ */
 @Component({
   selector: "app-product",
   templateUrl: "./product.component.html",
   styleUrls: ["./product.component.scss"],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ProductComponent implements OnInit {
-  private logger: string;
+export class ProductComponent {
+  /**
+   * Product's list input
+   */
+  @Input()
+  public productList: Array<ProductModel>;
 
   /**
-   * Event emitter for button
+   * Event emitter for on update cart button click
    */
   @Output()
-  public buyClickEvent: EventEmitter<string> = new EventEmitter();
+  public updateCart: EventEmitter<ProductModel> = new EventEmitter();
 
   /**
-   * ngOnInit
+   * On update cart button click
    */
-  public ngOnInit(): void {
-    this.logger = "Button was clicked";
-  }
-
-  /**
-   * On buy button click
-   */
-  public onBuyClick(): void {
-    this.buyClickEvent.emit(this.logger);
+  public onUpdateCart(product: ProductModel): void {
+    this.updateCart.emit(product);
   }
 }