+++ src/app/products/index.ts
@@ -1,3 +1,4 @@
 export * from "./components/index";
 export * from "./models/index";
 export * from "./services/index";
+export * from "./products.module";